Added code to copy the name for each argument into the LLVM IR.
Interestingly, the Rust version of the LLVM API is emulating a
subclass relationship on the single `LLVMValueRef` type by introducing
two distinct reference types `&Value` and `&Arg` that both map to
`LLVMValueRef`, and then a specialized `Deref` on `Arg` that converts
it to a `Value`.

(Unfortunately `&Value` and `&Arg` are references to empty structs
rather than a struct with private zero-sized field, which makes the
references forgeable. Maybe I will post a change to the API for that.)

(This same trick is applied to `&Function` in that same module.)